build(lint-staged): added lint-staged
On commit eslint was run with --fix. This would change files and these files would not be part of
the commit. Switched to lint-staged to only lint staged files and commit them, so that they are part
of the commit. Non staged files and untracked files are stashed before linting and unstaged ones
completed.